app/pyproject.toml
[tool.poetry]
name = "picpay-api"
version = "0.1.0"
description = ""
authors = ["Jonathan Bispo <jonathanbispo.pessoal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
loguru = "^0.7.2"
pydantic-settings = "^2.3.4"
joblib = "^1.4.2"
numpy = "^2.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.5"
mypy = "^1.11.0"
pytest = "^8.3.2"
locust = "^2.29.1"
asgi-lifespan = "^2.1.0"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# max number of characters per line
line-length = 88